requirements.txt
torch==1.11.0+cu113
torchvision==0.12.0
torch-fidelity==0.3.0

opencv-python==4.5.3.56
matplotlib==3.5.2
imageio==2.16.1

numpy==1.24.3
tqdm==4.66.1
nni==2.10

onnxruntime==1.10.0
scikit-learn==1.4.0
openvino==2022.3.0
pandas==2.2.0
Pillow==10.0.1
prettytable==3.9.0
protobuf==4.23.4
psutil==5.9.8
pytoolconfig==1.2.5
pyOpenSSL==23.2.0
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
pytorch-fid==0.3.0
pytorch-lightning==2.0.0
PyYAML==6.0.1
scikit-image~=0.22.0
scikit-learn==1.4.0
scipy==1.11.1
setuptools==68.2.2
tensorboard==2.15.1
tensorboard-data-server==0.7.2
